SoLazer fardo detector/requirements.txt
tensorflow
numpy
pillow
matplotlib
